Create vmd_mrt.sh
Analyses a NAMD trajectory to calculate the mean residence time of a solvent in an ions solvation shell.
Note! This needs work - I need to incorporate in a transition region in and out of the solvation layer. And it is currently written for some files that had 20ps per frame.